PEP8 Formatting and Commenting (excl Gamekeeper stuff)
Additionally added the citations for C02 stats and formatted the document in latex.

+++ technical-documents/source-code/gkHomepage/views.py
@@ -119,47 +119,47 @@ def index(request):
 
         #if someone presses the accept button
 
+        if ImageSubmission.objects.all().count() > 0:
 
 
+            if request.method == 'POST' and 'action_btn_accept' in request.POST:
+                print("----", "YOU'VE PRESSED ACCEPT")
+                images = ImageSubmission.objects.exclude(
+                    id=ImageSubmission.objects.first().id)
+                args = {'images': images}
 
-        if request.method == 'POST' and 'action_btn_accept' in request.POST:
-            print("----", "YOU'VE PRESSED ACCEPT")
-            images = ImageSubmission.objects.exclude(
-                id=ImageSubmission.objects.first().id)
-            args = {'images': images}
+                top_sub = get_top_submission()
 
-            top_sub = get_top_submission()
+                # Calculate statistics for user
 
-            # Calculate statistics for user
+                #get the username of the users imag
+                username = get_top_username()
 
-            #get the username of the users imag
-            username = get_top_username()
+                if top_sub == None:
+                    return render(request, "gkHomepage/gkHomepage.html", args)
+                username = top_sub.user
 
-            if top_sub == None:
+                user = CustomUser.objects.get(username=username)
+                #calulate the users streak(if any)
+                calc_user_streaks(user, datetime.today())
+
+                print("----",get_top_submission().building)
+                #calculate the points to give the user
+                points = calcPoints(get_top_submission().building)
+                #add the points to the users account
+                addPoints(username, points)
+                #remove that image from the database
+
+                print("----", top_sub.building)
+                points = calcPoints(top_sub.building)
+                addPoints(username, points)
+
+                # Checks if stats can be input and inputs if so
+                input_stats(top_sub)
+
+                ImageSubmission.objects.all().first().delete()
+                #render the template again, checking if theres a new image to upload
                 return render(request, "gkHomepage/gkHomepage.html", args)
-            username = top_sub.user
-
-            user = CustomUser.objects.get(username=username)
-            #calulate the users streak(if any)
-            calc_user_streaks(user, datetime.today())
-
-            print("----",get_top_submission().building)
-            #calculate the points to give the user
-            points = calcPoints(get_top_submission().building)
-            #add the points to the users account
-            addPoints(username, points)
-            #remove that image from the database
-
-            print("----", top_sub.building)
-            points = calcPoints(top_sub.building)
-            addPoints(username, points)
-
-            # Checks if stats can be input and inputs if so
-            input_stats(top_sub)
-
-            ImageSubmission.objects.all().first().delete()
-            #render the template again, checking if theres a new image to upload
-            return render(request, "gkHomepage/gkHomepage.html", args)
 
         #if the user presses the delete button
         if request.method == 'POST' and 'action_btn_delete' in request.POST:
@@ -184,7 +184,7 @@ def index(request):
             #get the username of the user who sumbitted the image
             username = get_top_username()
             user = CustomUser.objects.get(username=username)
-            image = stR(get_top_submission().image)
+            image = str(get_top_submission().image)
             print("----",image)
             #generate an email to send to the univeristy
             email = EmailMessage(
